clean up and separation of concerns
